requirements.txt
pyspark==3.5.1
delta-spark==3.2.0
pandas>=2.2
pyarrow>=16.0
python-dotenv>=1.0
jupyterlab>=4.2
ipykernel>=6.29